Clean up SearchSource and get rid of AbstractDataSourceProvider (#16932)
* Merge search source and abstract and es6ify SearchSource class

* Remove duplicate function and rename SourceAbstracts to SearchSource
